+++ actingweb/db/dynamodb/__init__.py
@@ -9,7 +9,7 @@ from .actor import Actor, CreatorIndex, DbActor, DbActorList
 from .attribute import Attribute, DbAttribute, DbAttributeBucketList
 from .peertrustee import DbPeerTrustee, DbPeerTrusteeList, PeerTrustee
 from .property import DbProperty, DbPropertyList, Property
-from .property_lookup import DbPropertyLookup, PropertyLookup
+from .property_lookup import DbPropertyLookup, PropertyLookup, PropertyLookupV2
 from .subscription import DbSubscription, DbSubscriptionList, Subscription
 from .subscription_diff import (
     DbSubscriptionDiff,
@@ -38,6 +38,7 @@ __all__ = [
     "DbPropertyList",
     # PropertyLookup
     "PropertyLookup",
+    "PropertyLookupV2",
     "DbPropertyLookup",
     # Subscription
     "Subscription",